pyproject.toml
[project]
name = "fact-check"
version = "0.1.0"
description = "This is a project for fact checking"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=0.1.0",
    "boto3>=1.28.0",
    "fastapi>=0.100.0",
    "httpx>=0.24.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "starlette>=0.27.0",
    "uvicorn>=0.23.0",
]
